fix spacec, fix folder's directories
