--- src/renderer/components/ComingSoonPages/MainPage.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const MainPage = () => {
-    return (
-        <div>Coming Soon...</div>
-    )
-}
-
-export default MainPage;
