--- src/renderer/components/ComingSoonPages/SettingsPage.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const SettingsPage = () => {
-    return (
-        <div>Settings are Coming Soon...</div>
-    )
-};
-
-export default SettingsPage;
