--- src/renderer/components/Pages/TransactionsPage.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import TransactionsList from "../TransactionsList";
-import { TransactionProvider } from "renderer/TransactionsContext";
-
-const TransactionsPage = () => {
-    return (
-        <div>
-        <TransactionProvider>
-         <TransactionsList />
-       </TransactionProvider>
-        </div>
-    )
-};
-
-export default TransactionsPage;
