+++ src/renderer/pages/MainPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MainPage = () => {
+  return <div>Coming Soon...</div>;
+};
+
+export default MainPage;
