+++ src/renderer/pages/SettingsPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const SettingsPage = () => {
+  return <div>Settings are Coming Soon...</div>;
+};
+
+export default SettingsPage;
